moved some helper methods to separate file versions_helper.rb to resolve cop failure
